登陆注册、试写WeChat Module

+++ app/src/main/java/xyz/sealynn/androidfun/APP.java
@@ -9,6 +9,7 @@ import com.squareup.leakcanary.LeakCanary;
 
 import androidx.appcompat.app.AppCompatDelegate;
 import xyz.sealynn.androidfun.utils.NightModeUtils;
+import xyz.sealynn.androidfun.utils.ToastUtils;
 
 /**
  * Created by SeaLynn0 on 2018/8/28 19:56
@@ -18,7 +19,8 @@ import xyz.sealynn.androidfun.utils.NightModeUtils;
 public class APP extends Application {
 
     private static Context appContext;
-//    private static long exitTime = 0;
+
+    private static long exitTime = 0;
 
     /**
      * 获取Application的Context
@@ -42,23 +44,18 @@ public class APP extends Application {
         Logger.addLogAdapter(new AndroidLogAdapter());
         appContext = getApplicationContext();
         AppCompatDelegate.setDefaultNightMode(NightModeUtils.getNightModeState(getAppContext()));
-//        if (NightModeUtils.getNightModeState(getAppContext())){
-//            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-//        }else {
-//            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-//        }
     }
 
     /**
      * 退出APP
      */
     public static void exitApp() {
-//        if (System.currentTimeMillis() - exitTime > 2000) {
-//            ToastUtils.shortToast(getAppContext(), appContext.getString(R.string.text_press_again));
-//            exitTime = System.currentTimeMillis();
-//        } else {
-//            android.os.Process.killProcess(android.os.Process.myPid());
-//        }
-        android.os.Process.killProcess(android.os.Process.myPid());
+        if (System.currentTimeMillis() - exitTime > 2000) {
+            ToastUtils.shortToast(getAppContext(), appContext.getString(R.string.text_press_again));
+            exitTime = System.currentTimeMillis();
+        } else {
+            android.os.Process.killProcess(android.os.Process.myPid());
+        }
+//        android.os.Process.killProcess(android.os.Process.myPid());
     }
 }
